// SGUEES-SPA/src/app/pages/Accounting/con-periodo-contable/models/con-periodo-contable.ts
export interface ConPeriodoContable {
	CORR_EMPRESA: number;
	ANIO_PERIODO: number;
	MES_PERIODO: number;
	NOMBRE_MES_PERIODO: string;
	ESTADO_PERIODO_CON: string;
	NOMBRE_ESTADO_PERIODO_CON: string;
	ESTADO_PERIODO_BAN: string;
	NOMBRE_ESTADO_PERIODO_BAN: string;
	ESTADO_PERIODO_VEN: string;
	NOMBRE_ESTADO_PERIODO_VEN: string;
	ESTADO_PERIODO_ACT: string;
	NOMBRE_ESTADO_PERIODO_ACT: string;
	ESTADO_PERIODO_INV: string;
	NOMBRE_ESTADO_PERIODO_INV: string;
	ESTADO_PERIODO_PLA: string;
	NOMBRE_ESTADO_PERIODO_PLA: string;
	ESTADO_PERIODO_COM: string;
	NOMBRE_ESTADO_PERIODO_COM: string;
	FECHA_CIERRE_CON: Date;
	FECHA_CIERRE_BAN: Date;
	FECHA_CIERRE_VEN: Date;
	FECHA_CIERRE_ACT: Date;
	FECHA_CIERRE_INV: Date;
	FECHA_CIERRE_PLA: Date;
	FECHA_CIERRE_COM: Date;
}